[PATCH] rework of ecryptfs home confidential directory setup
This patch represents a total rework of the helper script associated
with this how-to:
 * http://ecryptfs.sourceforge.net/ecryptfs-pam-doc.txt

CHANGES:

 * src/utils/Makefile.am: Updated makefile, I'm really not sure if I did
this correctly, please review these changes closely.  Please feel free
to update this accordingly upon checkin if necessary.

 * src/utils/ecryptfs-mount-confidential: Small script extracted from
ecryptfs-setup-pam-wrapped.sh.  This is the code that previously was
appended to the user's bash scripts.  Instead, create a standalone
script, runnable by any user to do this.

 * src/utils/ecryptfs-setup-confidential: This script is the rework of
ecryptfs-setup-pam-wrapped.sh.  The changes include:
  - header added, with description, history, and at least a Canonical
copyright; Mike can add an IBM one if necessary
  - created a usage() function, parameter description updated
accordingly, exits 1
  - use consistent indentation (tabs)
  - removed all known bashisms, so that this operates in a
posix-compatible shell (eg, dash)
  - check for username in /etc/passwd
  - look for mount/login password first on the command line, then as an
environment variable, and finally if both are empty, interactively
prompt
  - look for /etc/pam.d/system-auth and then /etc/pam.d/common-auth;
this provides support for both RH-based and Debian-based PAM-setups
  - collapse the mkdir/chown operations to a single command line
  - use mktemp for tempfiles
  - strengthen grep regex's
  - use chmod/chown --reference when overwriting config files
  - rebuild the pam conf file in perhaps a more reliable manner
  - add a call to the ecryptfs-mount-confidential script
to .bash_profile
  - add an unmount call to .bash_logout
  - backup any previous wrapped-passphrase file to a timestamped rename
 * src/utils/ecryptfs-setup-pam-wrapped.sh: deleted

TODO:

 * There are two BUG's noted in src/utils/ecryptfs-setup-confidential,
where passphrases will be momentarily visible in the process table to
unprivileged users.  A viable workaround should be established for
these.

TESTING:
 * I've tested this thoroughly under Ubuntu Hardy.

Signed-off-by: Dustin Kirkland <kirkland@canonical.com>