ecryptfs-setup-confidential: default to random mount passphrase
Per lengthy discussions on IRC [1], this patch changes the behavior such
that interactive prompting for a mount passphrase will allow a user to
just hit enter and have a random one generated (128 bits of random data
from /dev/urandom, converted into a hex string).

An accentuated message will be printed to standard out, informing the
user that they should record that passphrase somewhere secure, and that
losing that passphrase could yield their encrypted data inaccessible
later.

I did not add a command line option for randomly generating the
passphrase.  I think it's reasonable for any wrapping utilities (such as
adduser) to generate the random passphrase (by the same or different
mechanisms) and pass that in as a parameter.

[1] http://irclogs.ubuntu.com/2008/06/10/%23ubuntu-server.html

Signed-off-by: Dustin Kirkland <kirkland@canonical.com>