ecryptfs-setup-confidential: rework to use mount.ecryptfs_private
This patch reworks the ecryptfs-setup-confidential in order to use the
new functionality and methodology possible with the
mount.ecryptfs_private and umount.ecryptfs_private setuid binaries.

The goal is to make the ecryptfs-setup-confidential script completely
executable by non-privileged users.

This patch makes the following changes:

 * Remove support for specifying non-standard MOUNTPOINT and CRYPTDIR
directories.  Set PRIVATE_DIR="Private".  This is a temporary measure.
However, the setuid mount.ecryptfs_private program and this script need
to be in sync.  I think this should configurable by a system
administrator by setting something in /etc/ecryptfs.conf (or the like).

 * Create an error_testing() function that cleans up cruft created by
the test scenarios.

 * Put the interactive user name prompting in a loop until a good
username is received.

 * Check for active mount points in /proc/mounts.  Look for both the
MOUNTPOINT and CRYPTDIR.  Error out if either is mounted.

 * Put login and mount passphrase interactive prompting in a loop until
proper input is received.  Use "stty" to hide input while the user is
typing.

  * Remove the PAM library name detection.  That was a nasty hack to fix
a bug where the build system was producing the wrong PAM library name.
Good riddance!

 * Remove the modprobe of ecryptfs.  This should happen automatically,
or by the packing/installer if at all.  Another privileged operation
that does not belong in this script.

 * Remove all of the /etc/fstab munging.  This is completely avoidable
now with the mount.ecryptfs_private and the umount.ecryptfs_private
setuid executables.  The /etc/fstab munging was messy and required root
privileges.  Good riddance to this too!

﻿* Remove the PAM configuration hack.  This is not permissible by Debian
policy, and requires system administrator privileges.  This only needs
to happen once per system, rather than every time a user sets up their
private directory.  This needs to be moved elsewhere.  Perhaps in a root
script, or perhaps in packaging.  But it does not belong here.

 * Remove the bash_proflie, bash_logout, and gnome autostart hacks.
These were temporary measures until we found better alternatives.  The
mount/unmount functionality will be handled by the ecryptfs PAM module.
That code is coming soon.  Stay tuned.

 * Backup both previous .ecryptfs/wrapped-passphrase
and .ecryptfs/Private.sig files.  Append the same timestamp to each (to
keep them paired for data recovery purposes).

 * Now call ecryptfs-wrap-passphrase and ecryptfs-add-passphrase
subsequently to write the wrapped passphrase to file, and add the
passphrase to the keyring in the current session.  We also grep/sed the
signature out of the ecryptfs-add-passphrase stdout and write that to
the .sig file.

 * Get rid of a bunch of various chown/chmod's that were necessary when
we were root running this on behalf of another user, or we were
sudo'ing.

 * Finally, we add a pretty rudimentary test scenario, wherein we:
  - mount the ecryptfs private directory we just created
  - grab roughly 16K (multiple pages) of random data
  - write that to a tempfile on the mountpoint
  - md5sum (read) that file, save to a variable
  - unmount
  - mount again
  - md5sum (read) that file again
  - compare the 2 different md5sums
  - error out if ANY of the above is not successful
 This test could be improved in various ways.  At some point, it should
probably be broken out to a script of its own.  For now, these
operations are very fast and perhaps can throw meaningful error messages
if the initial mount/write/unmount/read fail.

Signed-off-by: Dustin Kirkland <kirkland@canonical.com>