As part of the Ubuntu Intrepid integration of Ecryptfs, I have created a number of manpages where none existed before.  In some cases, this is for new utilities that I've written.  In other cases, this is for other ecryptfs-* utilities that exist but without documentation.
I had to "wing it" in a few spots.  Thus, I'd appreciate a careful
review.  I erred on the side of caution, not saying much at all where I
lacked an understanding (see *zombie*).  However, I think this is much
better than before, as now we at least have a framework of a manpage in
place, with the bulk of the groff formatting working.

Contributed by Dustin Kirkland.