.ecryptfs/auto-umount support
This very simple patch adds support for ~/.ecryptfs/auto-umount.

This allows a user to separately configure automatic mounting and
unmounting of ecryptfs directories.

It is conceivable that a user might want to automatically mount their
ecryptfs directory, but NOT unmount it (background processes, and such).

Conversely, a more conservative user might NOT want to mount their
ecryptfs directory automatically, but always want to automatically
unmount it.

Dustin Kirkland