pam_ecryptfs support for auto-mount and auto-umount
The ecryptfs-setup-private script touches .ecryptfs/auto-mount and
.ecryptfs/auto-umount.  Since the pam module is now actually
performing the mount, it needs to respect the user's configuration
specified by the existence of those two files.

This simple patch adds support for that.  I have tested it thoroughly
and we are carrying the patch in Ubuntu Intrepid.

Eventually, I'd like to move the naming of those to files to something
more like:
 * .ecryptfs/Private.auto-mount
 * .ecryptfs/Private.auto-umount
files containing the name of that which is to be mounted/unmounted.

However, we'd need to do something for legacy support and the
appropriate renaming for existing users.  That's too ambitious at this
point, and we should probably consider a richer configuration file
before going that route.

In any case, this simple patch is very useful.  Please apply.