Improve "default salt" warning message
Multiple Ubuntu users have complained about the confusing
nature of the current "default salt" warning message.

Improve this message from:
  Unable to read salt value from user's .ecryptfsrc file; using default
to
  Warning: Using default salt value

This clearly states that this is a warning.  Also, the location
of the .ecryptfsrc should be handled in documentation, not
this error message.

Since this message is used multiple times, put it into a #define
in ecryptfs.h.  This will help with translatability (someday).

Finally, these messages should be printed on stderr, not stdout.

Signed-off-by: Dustin Kirkland <kirkland@canonical.com>