ecryptfs-setup-private should limit password attempts
ecryptfs-setup-private should limit the number of login
password attempts. Currently a user needs to hit Ctrl-c
to exit from the password prompt loop.

It is common practice (eg login) to limit the number of
attempts (though this is not a security issue-- anyone
could script up something for unix_chkpwd).

https://bugs.launchpad.net/bugs/291227

Signed-off-by: Dustin Kirkland <kirkland@canonical.com>