allow non-login wrapping passphrases
If a user is not using PAM, they may want to set up an encrypted Private
directory using a wrapping passphrase different from their login
passphrase.

Allow for this with a --wrapping option to ecryptfs-setup-private, which
will touch ~/.ecryptfs/wrapping-independent.  Have
ecryptfs-mount-private check for the existence of this file, and if so,
prompt for a wrapping passphrase rather than a login passphrase.

Signed-off-by: Dustin Kirkland <kirkland@canonical.com>