mount.ecryptfs_private: user chosen mountpoint
Numerous users have requested more flexibility in the naming of the
mountpoint of their ~/Private directory, for various reasons (mainly for
translatability reasons).

This patch adds support for a user specified mountpoint in
~/.ecryptfs/Private.mnt.

mount.ecryptfs_private will ensure that the chosen mountpoint exists, is
a directory, and is owned by the user.

This patch has the extremely beneficial side-effect that it **vastly**
simplifies the encrypted home directory code.  We can git rid of all of
the special-case hack for encrypting all of home.  The user can simply
`echo $HOME > ~/.ecryptfs/Private.mnt`.

Signed-off-by: Dustin Kirkland <kirkland@canonical.com>