mount.ecryptfs_private: rework mount counter
This is a more complete fix for this bug:
https://bugs.edge.launchpad.net/ubuntu/+source/ecryptfs-utils/+bug/304043

This patch completely re-works the mount counter, establishing a
comprehensive lock on the counter file for the length of the program's
run.

Also, silence compiler warnings with GNU_SOURCE, and values.h.

Finally, create fail and success goto's, for cleaner exits.

Signed-off-by: Dustin Kirkland <kirkland@canonical.com>