lower fnek iterations
The 65K iterations performed to generate a key signature is intended
to make dictionary attacks more expensive.

The FNEK is itself a hash (16 arbitrary bytes of data), and therefore
not susceptible to dictionary attacks, and thus these iterations are not
so necessary.  Let's lower it from 65K to 1024.  Leaving it there for
good measure?

Signed-off-by: Dustin Kirkland <kirkland@canonical.com>