move the modprobe to a better place
Move the modprobe ecryptfs code to the get-version function.
Use /sbin since it might not be in the path.
Discard errors, since this code might not be called by root.
Add a comment, explaining what's going on.

A printf() if it still fails, noting that the user might need
to run modprobe ecryptfs as root might be a nice enhancement.

From: Michal Hlavinka <mhlavink@redhat.com>
Signed-off-by: Dustin Kirkland <kirkland@canonical.com>