Automatically unlink fekek and fnek at unmount with mount helpers
A feature request was made for the automatic removal of the fekek and
fnek from the keyring at unmount time when using the mount helpers.
The best way to do this involved creating an eCryptfs umount helper,
but I didn't want it to affect the users who don't use the eCryptfs
mount helper.

This patch creates a new eCryptfs umount helper that looks in the mount
options for ecryptfs_unlink_sigs.  If that option is found, it will
determine the ecryptfs_sig and the ecryptfs_fnek_sig, if one exists, and
unlink those auth toks from the user keyring.  If the
ecryptfs_unlink_sigs option is not detected, a normal umount will occur.

By default, the mount helper will now insert ecryptfs_unlink_sigs into
the mount options.  This will not affect users who have pre-existing 
fstab entries and are not expecting their keys to unlinked at umount.

Signed-off-by: Tyler Hicks <tyhicks@linux.vnet.ibm.com>