Ignore unknown kernel ciphers
libecryptfs stores a list of known ciphers that are either whitelisted
or blacklisted for use with eCryptfs mounts.  It parses /proc/crypto to
determine the loaded kernel ciphers and scrubs any that are blacklisted.
However, if there are loaded ciphers that are unknown (not whitelisted
or blacklisted), they are mistakenly allowed into the whitelisted
ciphers.

This patch ignores any unknown ciphers and also blacklists des3_ede128.

Fixes: Bug #335632 - unknown ciphers can be selected in mount helper

Signed-off-by: Tyler Hicks <tyhicks@linux.vnet.ibm.com>