pam_ecryptfs: check ecryptfs support for kernel module loading
When auto-mounting, ecryptfs kernel module does not have to be
in kernel yet. ecryptfs_insert_wrapped_passphrase_into_keyring
call checks if kernel supports file name encryption, but because
this module is not loaded and because it can't load kernel module
itself since called with uid of the user, file name encryption
is not used.

Signed-off-by: Michal Hlavinka <mhlavink@redhat.com>