build: use pkg-config to identify presence of NSS.
The nss-config script is not cross-compile aware, but pkg-config is (or can
be made so). Instead of calling directly the former, use the latter.

This also allows for the flags and libs variables to be overridden by the
user at ./configure time, like any other important variable, and decouples
from nspr dependency, as nss pkg-config file already imports that.