tests/lib/etl_funcs.sh: workaround XFS truncate issues
XFS runs really sub-optimally when using small images with single files
and large truncate operations. This is a known performance characteristic
with XFS and so we should avoid almost filling up the file system. For
small partitions, use just 1/3 of the file system, otherwise just use 1/2
to avoid sucky slow performance with XFS on the loop-back.