tests/kernel/mmap-bmap: Add bmap mmap test
eCryptfs bmap is not yet being exercised by the eCryptfs tests so add
a simple FIBMAP test that checks to see if an encrypted file blocks are
contained in the lower file system.  For lower file systems that don't
support FIBMAP this test is effectively always a no-op.  Checked against
kernel gcov to see if bmap is being exercised.